+++ lib/measure.dart
@@ -1,10 +1,10 @@
 
 library measure;
 
-import 'dart:math' as Math;
+import 'dart:math' as math;
 import 'dart:collection';
 import 'package:quiver/core.dart' as quiver;
-
+import 'package:dart2_constant/math.dart' as math;
 export 'src/unit/unit_format.dart';
 
 part 'src/quantity.dart';